+++ lib/l10n/app_localizations_pt.dart
@@ -710,7 +710,13 @@ class AppLocalizationsPt extends AppLocalizations {
   String get prognosisHorizonLabel => 'Horizon';
 
   @override
+  String get prognosisHorizonTwoWeeks => '2 weeks';
+
+  @override
   String get prognosisHorizonEndOfMonth => 'End of month';
+
+  @override
+  String get prognosisHorizonMidNextMonth => '1.5 months';
 
   @override
   String get prognosisHorizonEndOfNextMonth => 'Next month';
@@ -735,6 +741,14 @@ class AppLocalizationsPt extends AppLocalizations {
 
   @override
   String get prognosisHorizonTenYears => '10 years';
+
+  @override
+  String get prognosisHorizonCustomDate => 'Pick a date';
+
+  @override
+  String prognosisHorizonUntil(String date) {
+    return 'Until $date';
+  }
 
   @override
   String get prognosisMilestoneThreeMonths => 'End of 3 months';